Fix TIMESTAMP and FLOAT type convertion

+++ flink-cdc-connect/flink-cdc-pipeline-connectors/flink-cdc-pipeline-connector-mysql/src/test/resources/ddl/inventory.sql
@@ -20,7 +20,7 @@ CREATE TABLE products (
   id INTEGER NOT NULL AUTO_INCREMENT PRIMARY KEY,
   name VARCHAR(255) NOT NULL DEFAULT 'flink',
   description VARCHAR(512),
-  weight FLOAT
+  weight FLOAT(6)
 );
 ALTER TABLE products AUTO_INCREMENT = 101;
